refactor(tmdb): extract types from tmdb-api.ts into separate types.ts
- Create new src/utils/types.ts with all TMDB type definitions
- Update tmdb-api.ts to import types from types.ts instead of defining inline
- Update component imports to use types.ts for MediaListItem type
- Improve code organization by separating type definitions from implementation
